fix: Bugfixes und Code-Qualitätsverbesserungen
- inputCheck.js: Logikfehler behoben (|| → &&), verhindert Runtime-Error bei leeren Feldern
- _app.js: Tippfehler im aria-label korrigiert ("Srollen" → "Scrollen")
- demoDataHelpers.js: createFileFromImage und createFileFromPDF in zentrale Utility-Datei ausgelagert
- Registrierungsformulare (8 Dateien): Duplizierte Funktionsdefinitionen entfernt, Import aus demoDataHelpers hinzugefügt
- console.log-Aufrufe aus allen Anmeldeformularen und sendEmail.js entfernt
- .env.local.example: Dokumentation aller benötigten Umgebungsvariablen hinzugefügt
