+++ .env.local.example
@@ -1,0 +1,12 @@
+# Datenbankverbindung (MySQL)
+DB_HOST=localhost
+DB_USER=dein_db_benutzer
+DB_PASSWORD=dein_db_passwort
+DB_DATABASE=dein_datenbankname
+
+# E-Mail (Nodemailer via mail.yumekai.de, Port 587)
+DEFAULT_EMAIL=deine@email.de
+DEFAULT_PASS=dein_email_passwort
+
+# E-Mail-Passwort für info@yumekai.de
+EMAIL_PASS_INFO=passwort_fuer_info_adresse
